Rename 'adaptor' to 'adapter'
